% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{IA_remove_y}
\alias{IA_remove_y}
\title{remove the y axis.}
\usage{
IA_remove_y()
}
\description{
Remove line, ticks, text and title of the Y axis.
very convenient for kernel densities.
Add it to a ggplot function as the last theme adjustment.
}
\examples{
p <- ggplot(mpg, aes(x = hwy)) + geom_density()
p
p + IA_remove_y()
}
