% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{IA_s_wrap}
\alias{IA_s_wrap}
\title{wrap a long string into several lines}
\usage{
IA_s_wrap(x, w = 20)
}
\description{
create a line break in long string before printing into a plot.
}
\examples{
# Normally str_wrap() is for 80 characters. This is to much for plots.
# So, this function is a wrapper with default of 20 characters to use easly in
# ggplot.


p <-
dplyr::filter(mpg,str_length(model) > 10) \%>\%
ggplot(aes(as.factor(year), displ)) +
geom_point()

p + facet_wrap(~ model)
p + facet_wrap(~ IA_s_wrap(model, 10))
}
