% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{IA_save}
\alias{IA_save}
\title{ggave with png device and white background}
\usage{
IA_save(file, plot = last_plot(), device = png, bg = "white")
}
\description{
Yotvat found bugs in the ggsave function. 1st, the saving does not specity
a background color. 2nd, in the new version one needs to specify the device.
this function is a wrapper around the ggsave with the correct defults to save time.
}
\examples{
# enter a file_name_with_path.png to save
}
