% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nice_html_table}
\alias{nice_html_table}
\alias{IA_t1}
\alias{IA_t2}
\alias{IA_t3}
\title{Print nice tables into the html document}
\usage{
IA_t1(df, digits = 3)

IA_t2(df, wi = 800, hi = 400, digits = 3)

IA_t3(df, group_var = NULL, digits = 3, wi = NULL, hi = 500, ...)
}
\arguments{
\item{df}{is a data frame}

\item{digits}{the number of rounded digits in numbers}

\item{wi}{is the width of the table in pixels.}

\item{hi}{is the height of the table in pixels}

\item{group_var}{in the IA_t3 this is an option to group rows together
by variable category.}
}
\description{
Add it with the pipe "\%>\%" at the end of a data frame
}
\examples{
IA_t1(head(mtcars))
IA_t2(head(mtcars))
IA_t3(mpg)
IA_t3(mpg, group_var = 'manufacturer', digits = 2, hi = 800)
IA_t3(mpg, group_var = c('manufacturer', 'model'))
}
