% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{Scale_y_continues_number_versions}
\alias{Scale_y_continues_number_versions}
\alias{IA_gg_k}
\alias{IA_gg_m}
\alias{IA_gg_p}
\title{scale y continues versions}
\format{
An object of class \code{ScaleContinuousPosition} (inherits from \code{ScaleContinuous}, \code{Scale}, \code{ggproto}, \code{gg}) of length 19.

An object of class \code{ScaleContinuousPosition} (inherits from \code{ScaleContinuous}, \code{Scale}, \code{ggproto}, \code{gg}) of length 19.

An object of class \code{ScaleContinuousPosition} (inherits from \code{ScaleContinuous}, \code{Scale}, \code{ggproto}, \code{gg}) of length 19.
}
\usage{
IA_gg_k

IA_gg_m

IA_gg_p
}
\description{
add scale_y_continues for thousands, millions of percentage.
}
\examples{
# Add the name of the environment at the end of a ggplot call
# preceded by a plus "+" sign.

p <- ggplot(txhousing,aes(year, volume, fill = city)) +
 geom_bar(stat = "summary", fun = "mean", show.legend = FALSE)

p
p + IA_gg_k
p + IA_gg_m


p <- ggplot(mtcars, aes(gear)) + geom_bar(aes( y = ..prop..))

p
p + IA_gg_p


}
\keyword{datasets}
